+++ src/variables/Variables.jsx
@@ -609,7 +609,7 @@ var legendBar = {
 module.exports = {
   style, // For notifications (App container and Notifications view)
   thArray,
-  tdArray, // For tables (TableList view)
+  tdArray, // For tables (Usuarios view)
   iconsArray, // For icons (Icons view)
   dataPie,
   legendPie,