+++ CHANGELOG.md
@@ -7,8 +7,8 @@
 - Renamed `src/views/Dashboard/Dashboard.jsx` to `src/views/Icons.jsx`
 - Renamed `src/views/Dashboard/Dashboard.jsx` to `src/views/Maps.jsx`
 - Renamed `src/views/Dashboard/Dashboard.jsx` to `src/views/Notifications.jsx`
-- Renamed `src/views/Dashboard/Dashboard.jsx` to `src/views/TableList.jsx`
-- Renamed `src/views/Dashboard/Dashboard.jsx` to `src/views/Typography.jsx`
+- Renamed `src/views/Dashboard/Dashboard.jsx` to `src/views/Usuarios.jsx`
+- Renamed `src/views/Dashboard/Dashboard.jsx` to `src/views/Noticias.jsx`
 - Renamed `src/views/Dashboard/Dashboard.jsx` to `src/views/Upgrade.jsx`
 - Renamed `src/views/Dashboard/Dashboard.jsx` to `src/views/UserProfile.jsx`
 - Renamed `src/components/Header` to `src/components/Navbars`